Write tests for app
Added a project to the directory so I could run django tests against it. I had to change some settings
in django to have an app that is outside of the project dir.

Wrote tests that should have 100% coverage of the project.

Mocked two tests that were making calls to recaptcha

Add test for testing the _original attribute function

Build 1.10

Add travis.yml file for build tests

try build

try build

try build

Rearranged the project directory to be on the same level as the app

try build

Build passing