Write tests for app
Added a project to the directory so I could run django tests against it. I had to change some settings
in django to have an app that is outside of the project dir.

Wrote tests that should have 100% coverage of the project.

Mocked two tests that were making calls to recaptcha